+++ src/ifcwrap/Interface.h
@@ -27,10 +27,10 @@ namespace IfcGeomObjects {
 	class IfcMesh {
 	public:
 		int id;
-		std::vector<float> verts;
+		std::vector<double> verts;
 		std::vector<int> faces;
 		std::vector<int> edges;
-		std::vector<float> normals;
+		std::vector<double> normals;
 		std::string brep_data;
 	};
 
@@ -41,7 +41,7 @@ namespace IfcGeomObjects {
 		std::string name;
 		std::string type;
 		std::string guid;
-		std::vector<float> matrix;
+		std::vector<double> matrix;
 		const std::vector<int> name_as_intvector() {
 			std::vector<int> r;
 			for ( std::string::const_iterator it = name.begin(); it != name.end(); ++ it ) r.push_back(*it);
